Added Class myDirectory Changed all critical System.IO.Directory-References to myDirectory. Now no DirectoryNotFoundException can occur, as long as the Directory is a subdirectory of Application.StartupPath.
Todo: Implement Functions to get video- and imagefiles

Description:
This Class is an Extension for the Class IO.Directory based on the needs of Tease-AI.
Exposes static methods for creating, moving, and enumerating through directories and subdirectories. This class cannot be inherited.
Implemeted this far:
- GetFiles() + 2 Overloads.
IO.Directory.GetFiles throws a Directory not FoundException, if the Directory of the file cannot be found. This Functions create the Directory, as long as it is in die Application.StartupPath.
- GetDirectories() +2 Overloads.
 IO.Directory.GetDirectories throws a Directory not FoundException, if the Directory of the file cannot be found. This Functions create the Directory, as long as it is in die Application.StartupPath.