BigFix: DommeTag-APP, @DommeTag(), HoldEgdeTaunts, Tna-Fast
This commit is WIP and  not suitable for publication!

BugFix: DommeTagApp wasn't showing the correct Tags for a displayed image. Beware: Functions Add and RemoveDommeTag are currently deactivated. They have to be re-adjusted.  The path of the current shown image is now avaiable in Form1.mainpicturebox.ImageLocation. This Property is used to get the current image for CheckDommeTag, because access through the index is not possible in combination with @DommeTag(). For Usage of CheckDommeTag see its XML-Comment.
BugFix: TnAFastSlides_Tick: Booblist was empty at start. But previous GetTnAList was performed... I have no clue. Marked as Qnd-Bugfix.
BugFix: HoldEdgeTauntTimer_Tick didn't load the TxtFile, resulting in "Tease AI did not Return..."
Bugfix: GetDommeTag():
 - The Function doesn't set any global variables anymore.
 - The return type is now a String, representing the found ImagePath. If none is found, it will return an empty String.

Deprected Variable "DommeImageListCheck": This Variable isn't needed anymore. Its marked as deprected. Please remove this Variable, without loosing the ability to Resume a Session - if that's even possible.