Improvement ImageLoading Status: Class is working, Integration Ongoing.
Added Class BackgroundWorkerSyncable, basically a Wrapper around System.ComponentModel.BackgroundWorker.Addes members to wait for the Backgroundworker to complete in a Application.DoEvents Loop and for delayed triggering the RunWorkerCompleted-Event.

Description
Now Images can be loaded on a BackgroundThread, while the calling procedure stops, until the the image has been loaded. Even manually syncing is possible. Simply Call ShowImage(ImageUrl, False). This way the Calling procedure is not blocked until the Backgroundworker is finished. In this case the BackgroundWorker wont trigger the RunWorkerCompleted-Event until BWimageFetcher.WaitToFinish() is called. This way the calling procedure can work while the Backgroundworker loads and saves the File. But this currently not recommended, because Tease-AI could do something unsuspected, like one of the more than 20 Timers calling the Backgroundworker again, while it's busy. The BackgroundWorker will remain busy until the RunWorkerCompleted-Event has been executed! This would throw an InvalidOperation-Exception.

Possible Improvements:
Add Timeout for pending RunWorkerCompleted-Event