New Temporary FilterList
New Temporary FilterList Sub - The old one has been renamed to
FilterListBak. I did as much work as I could ()before passing out for
the night) to get this back on par with the original FilterList while
actually working like it's supposed to on multi-line taunts. This is a
WIP of a temporary solution, but it should be no worse than the original
FilterList. It should be better