BugFix: Commands @PornAllowedOff and @PornAllowedOn where not replaced correct in Domme Output.. Added: Function IsAvailable to ImageDataContainer to simply Check if a ImageGenre is in general available and has Images to display. Basically it creates the ImageList and checks if there are ImageLocations in it. There is still room for better Resource Management... Added: ImageDataContainer is now able to check the sys_NoPornAllowed flag. Removed: Obsolete Function CheckGenreImage(String) replaced by ImageDataContainer.IsAvailable()
Note: sys_NoPornAllowed is activated on the next line and not in the Line containing the CreateCommand.

#TODO-List:
Replace Function CheckLocalImage(String, Boolean). This Function is "strange". I guess this shows the Images through PBImageThread. The Plan is to remove this and merge with the Function CheckPreLoad(String). A little distracting is the fact CheckPreLoad(string) is called before the current DommeTask is calculated. This way possibly a wrong image could be displayed.