@CustomMode() Command and Mode Class, Tasks overhaul tweaks
Added @CustomMode() Command and Mode Class - You can use @CustomMode to
set any keyword or phrase to a mode that will move to a GotoLine in the
script when used. The current types that @CustomMode can use are Goto
and Video. For example:

@CustomMode(I'm done, Goto, Finished Task) - This would move the script
to the line (Finished Task) if the user types "I'm done" (not case
sensitive)
@CustomMode(I saw boobs, Video, Video Closed) - This would stop a video
and move to the line (Video Closed) if the user types "I saw boobs" (not
case sensitive)

Only one @CustomMode() Command should be used per line, but you may have
as many custom modes activated at a time that you wish. Remember that
individual Modes are cleared when they resolve, and all modes are
cleared when transitioning between script types. To manually clear a
mode, use @CustomMode(ModeText, Normal)