Improved @If[] Command
Cleaned up @If[] Code and added "And" and "Or" capability.

Improved @If[] Command - @If[] can now use "And" and "Or" when making
comparisons. For example:

@If[Var_EdgeCount]>[10]And[#DommeMood]>[@DomMoodMax]Then(GotoLine)
@If[Var_EdgeCount]<[3]Or[Var_TotalAttempts]>[5]Then(GotoLine)

@If[Var_EdgeCount]>[10]And[#DommeMood]>[@DomMoodMax]And[#CBTCockCount]<[1]Then(GotoLine)

You can use as many "And" or "Or" comparisons per @If[] Command that you
like, but you CANNOT use both "And" and "Or" in the same @If[] Command

Started working on this this morning while waiting to hear back on the
settings, figured I would go ahead and finish it.  Now back to the joys
of DataBindings