CustomSlideshow Rework Fixed issue where a domme images was displayed, when a contact was responding.
Added Class CustomSlideshow:
This Class is a Dictionary(of String, Imagegenre) which is extended with navigation methods. It also has a built in RandomImage function, which will most likely return pretty random genres. Instead of picking one of all images, it picks a random genre and afterwards a random image out of the genre. It also remembers the last picked genre and tries to find another random genre, for  "number of genres in list divided by 4" times. This way all genres have a chance to get picked no matter how pictures are in. It is also possible to enhance the chances to pick a local path. This is useful for fast slideshows, where downloads take longer as the actual display time. Currently it is hard coded for @Slideshow(...... , fast) and regular.

Moved NullRespose jumpmarks in Timer1 and SendTimer, to allow @NullResponse with image Commands.
Moved NullRespose Jumpmarks in Timer1 and SendTimer, to allow @Nullresponse with image Commands.