@Goto() Improvements
Previously, Tease AI would crash if it tried to find a Goto Label that
did not exist in the current script. The two most common occurrences
seemed to be when INterrupts were called while a Goto line was active,
and typos in the @Goto() Command or Goto Label. Goto searches are now
cleared when Interrupts, CallReturns or Miniscripts are used.
Additionally, if Tease AI cannot find an exact match after parsing the
entire script, it will begin checking labels for similar text found in
the @Goto() Command, looking for any Labels that are 1 character off,
then 2, then 3 and so on up to 5. This should help scripts flow
correctly even when typos are present. If Tease AI still fails to find
an exact or similar match, then the program will move on to a random
Link script and begin the next cycle. This should prevent the widespread
problem of crashes occurring when @Goto() can't find what it's looking
for.