New command: @DommeTagOr(): will choose a domme tagged image having at least one of the required tag. @DommeTag() will (now working properly) look for an image having ALL the required tags instead
Changes:
@DommeTag and @DommeTagOr now use the same functions used by ImageTag and ImageTagOr instead of the function getTaggedImage in the contactData class.
That function was not working properly (for example @DommeTag(Boobs,Naked) would never return anything since it was looking just for a single tag...and it would have considered Boobs,Naked as the tag to look for) and i think it was useless to have 2 functions basically doing the same stuff but for different commands.
P.s: the taggedImage function in contactdata is still there since i didn't want to cancel it in case you didn't agree with my changes...but as of now it is never used

Fixes:
Fixed a pair of errors in my "safenet" code for when the user forget  @End or @StartStroking/Taunts in a script.....i forgot to reset the isLink boolean to false when calling an @Interrupt().